Use base class UiSubobject to hold pointer to Ui
This makes construction of UI subobjects simpler, especially for the
UiFactory class.

The base class UiSubobject iss used to hold a non-owning pointer to the
overseeing Ui instance. Specifically, the referred object is not
destroyed upon destruction of the subclass object.

The classes Ui/UiFactory own their subcomponents and are responsible for
their destruction.