Add game loop to ui
Member function act() to perform a single action with the current
palyer. Member function startGame() to run act() until the game is over.
The UiFactory class is required to construct Ui objects in such a way
that the underlying game has started with the first player to move when
finalyzing the Ui object.