Add ui::Average subobject to UI
Call addToAverage() in the relay of Game_t::finishTurn().
Since the game currently does not return the obtained points in any way,
compute the turn points separately by querying the points of the moving
player before and after finishing her turn. The difference is the number
of points obtained during the turn.

In the future, the game class should probably return the turn points
itself.