Implement class "GameState" for game-style play
Class "GameState" (planned to later be renamed "Game") implements
incremental interactions with an "Environment" instance.

Internally, the class "GameState" stores an "Environment" object that
dictates outcomes of the randomized events, i.e., Initialization and
rolling of dice.
A "Cup" object is maintained to be consistent with the environment. The
position of each of the environment's dictated digit is randomized
within the cup, creating the illusion that the dictated random digits
were created by rolling each die individually.