Implement class "Tenthousand"
Class Tenthousand provides an interface for the interactive multiplayer
game until 10000 points. Available game actions are fully sanity
checked. Non-game-interactions, however, may *not* be sanity checked.
For example, one may not query the winner without ensuring that a winner
exists; and one may not enter actions if the game is not running
anymore.