Hide unusable dice on user interface
This feels more similar to a cup of dice where some dice are put aside
so they are no longer visible. If this system tests positive for human
interaction, it may be considered to move the remaining dice all the way
to the left (low indices), to easy recognition and typing.