Automate plot generation
Automate generation of gnuplot script to
  - plot most recent, temporary training data.
  - plot multiple previously stored sets of training data for
    comparison.

Plot generation would be desirable for any saved evaluation ID (not just
0). This would need adjustments to script the create_gnuplot_script.sh.
The evaluation ID must be baked into the plotscript.

Athough, thereis probably a way to make the gnuplot scrip adjust to the
desired ID dynamically, as long as the validity of the ID is ensured.